backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4
asyncpg
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
databases
redis
httpx
python-multipart
pandas
python-dotenv
